requirements.txt
aiohttp>=3.9.3
aiofiles>=23.2.1
colored>=2.2.4
croterline>=1.0.3
GitPython>=3.1.43
httpx>=0.27.0
nonebot-plugin-htmlrender>=0.1.0
nonebot2[fastapi,httpx,websockets]>=2.3.3
nonebot-adapter-onebot>=2.4.3
nonebot-plugin-alconna>=0.46.3
nonebot_plugin_apscheduler>=0.4.0
nonebot-adapter-satori>=0.11.5
pyppeteer>=2.0.0
markdown>=3.3.6
zhDateTime>=1.0.3
numpy>=2.1.0
packaging>=23.1
psutil>=5.9.8
py-cpuinfo>=9.0.0
Pygments>=2.17.2
pyppeteer>=2.0.0
pytz>=2024.1
PyYAML>=6.0.1
pillow>=10.0.0
toml>=0.10.2
importlib_metadata>=7.0.2
watchdog>=4.0.0
jieba>=0.42.1
python-dotenv>=1.0.1
loguru~=0.7.2
pydantic~=2.9.2
pip~=23.2.1
fastapi~=0.115.0
magicoca~=1.0.1